+++ src/pyosmeta/__about__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: 2023-present Leah Wasser <leah@pyopensci.org>
-#
-# SPDX-License-Identifier: MIT
-__version__ = "0.0.1"