+++ src/pyosmeta/parse_issues.py
@@ -265,6 +265,51 @@ class ReviewModel(BaseModel):
 
         return user
 
+    @field_validator(
+        "categories",
+        mode="before",
+    )
+    @classmethod
+    def clean_categories(cls, categories: list[str]) -> list[str]:
+        """Make sure each category in the list is a valid value.
+
+        Valid pyos software categories are:
+            citation-management-bibliometrics, data-deposition,
+            data-extraction, data-processing-munging, data-retrieval,
+            data-validation-testing, data-visualization-analysis,
+            database-interoperability, education,
+            geospatial, scientific-software-wrappers,
+            workflow-automation-versioning
+
+
+        Parameters
+        ----------
+        categories : list[str]
+            List of categories to clean.
+
+        Returns
+        -------
+        list[str]
+            List of cleaned categories.
+        """
+
+        valid_categories = {
+            "data-processing": "data-processing-munging",
+            "data-validation": "data-validation-testing",
+            "scientific-software": "scientific-software-wrapper",
+        }
+
+        cleaned_cats = []
+        for category in categories:
+            for valid_prefix, valid_cat in valid_categories.items():
+                if category.startswith(valid_prefix):
+                    cleaned_cats.append(valid_cat)
+                    break
+            else:
+                # No match found, keep the original category
+                cleaned_cats.append(category)
+        return cleaned_cats
+
 
 @dataclass
 class ProcessIssues: